+++ script/core/highlight.lua
@@ -267,7 +267,7 @@ return function (uri, offset)
             if isLiteral ~= isLiteralValue(target) then
                 return
             end
-            if not files.eq(uri, guide.getUri(target)) then
+            if uri ~= guide.getUri(target) then
                 return
             end
             local kind