+++ script/core/type-definition.lua
@@ -120,7 +120,7 @@ return function (uri, offset)
     if uris then
         for i, uri in ipairs(uris) do
             results[#results+1] = {
-                uri    = files.getOriginUri(uri),
+                uri    = uri,
                 source = source,
                 target = {
                     start  = 0,
@@ -160,7 +160,7 @@ return function (uri, offset)
         or src.type == 'doc.type.ltable' then
             results[#results+1] = {
                 target = src,
-                uri    = files.getOriginUri(root.uri),
+                uri    = root.uri,
                 source = source,
             }
         end